Initial commit
This is just a really basic implementation of collocation with a
monomial basis and a hardcoded choice between uniform grid and Legendre
roots (which are hardcoded for 0 to 1 instead of t0 to tf).

The uniform grid works better than the Legendre roots, but overall the
solver doesn't seem very good, as it can claim convergence even when the
initial value is way off.

Using jacfwd/jacrev led to some improvement, but also we sometimes get
really weird polynomils from the uniform grid, despite claims of
convergence.