A few error have been corrected, and I now have it working for generic t0-tf
It's struggling mightily on a larger domain, and so I think adding
support for "finite elements" is necessary in order to make this work on
larger domains. I could also probably use a better representation, but
I'm curious as to how far this naive one can go and this exercise is
interesting for showing the various pitfalls that this approach can fall
into.